Refactor autoProcess -> auto_process
