+++ core/auto_process/__init__.py
