+++ core/auto_process/autoProcessComics.py
