+++ core/auto_process/autoProcessGames.py
